server/apps/core/api/urls.py
from django.urls import include, path
from rest_framework import routers

from . import views

# Routers provide an easy way of automatically determining the URL conf.
home_router = routers.DefaultRouter()
home_router.register(r"books", views.BookViewSet)

urlpatterns = [
    path("", include(home_router.urls), name="api-home"),
    path("csrf/", views.get_csrf, name="api-csrf"),
    path("login/", views.login_view, name="api-login"),
    path("logout/", views.logout_view, name="api-logout"),
    path("session/", views.session_view, name="api-session"),
    path("whoami/", views.whoami_view, name="api-whoami"),
]
